Minor and major changes to the certain pages in the app router - Converted "Add Students" and "Add Supervisors" pages back to server components. - Renamed "Algorithms Overview" to "Algorithms". - Renamed "Algorithm Details" page to "Results". - Excluded export to external system page from app router.
