+++ src/app/(protected)/[group]/[subGroup]/[instance]/(admin-panel)/(stage-4)/algorithms/_components/new-algorithm-button.tsx
