+++ src/app/(protected)/[group]/[subGroup]/[instance]/(admin-panel)/(stage-4)/algorithms/_components/results-table.tsx
