+++ src/app/(protected)/[group]/[subGroup]/[instance]/(admin-panel)/(stage-4)/results/_components/details-data-table.tsx
