+++ src/app/(protected)/[group]/[subGroup]/[instance]/(admin-panel)/(stage-4)/results/page.tsx
@@ -5,7 +5,7 @@ import { api } from "@/lib/trpc/server";
 import { InstanceParams } from "@/lib/validations/params";
 import { adminTabs } from "@/lib/validations/tabs/admin-panel";
 
-import { AddSupervisorsSection } from "./_components/add-supervisors-section";
+import { ResultsSection } from "./_components/results-section";
 
 import { app, metadataTitle } from "@/content/config/app";
 import { pages } from "@/content/pages";
@@ -14,15 +14,19 @@ export async function generateMetadata({ params }: { params: InstanceParams }) {
   const { displayName } = await api.institution.instance.get({ params });
 
   return {
-    title: metadataTitle([pages.addSupervisors.title, displayName, app.name]),
+    title: metadataTitle([pages.results.title, displayName, app.name]),
   };
 }
 
 export default function Page() {
   return (
-    <PanelWrapper className="mt-10">
-      <SubHeading>{adminTabs.addSupervisors.title}</SubHeading>
-      <AddSupervisorsSection />
+    <PanelWrapper className="mt-20 flex flex-col items-center">
+      <div className="flex w-full flex-col gap-3">
+        <SubHeading className="mb-6 text-2xl">
+          {adminTabs.algorithmDetails.title}
+        </SubHeading>
+        <ResultsSection />
+      </div>
     </PanelWrapper>
   );
 }
