+++ src/app/(protected)/[group]/[subGroup]/[instance]/(admin-panel)/(stage-6)/_export-to-external-system/_components/check-users-section.tsx
