+++ src/app/(protected)/[group]/[subGroup]/[instance]/(admin-panel)/(stage-6)/_export-to-external-system/_components/send-allocations.tsx
