+++ src/app/(protected)/[group]/[subGroup]/[instance]/(admin-panel)/(stage-6)/_export-to-external-system/page.tsx
@@ -1,7 +1,11 @@
-import { Unauthorised } from "@/components/unauthorised";
+import { SubHeading } from "@/components/heading";
+import { PanelWrapper } from "@/components/panel-wrapper";
+import { NoteCard } from "@/components/ui/note-card";
 
 import { api } from "@/lib/trpc/server";
 import { InstanceParams } from "@/lib/validations/params";
+
+import { CreateProjectsSection } from "./_components/send-allocations";
 
 import { app, metadataTitle } from "@/content/config/app";
 import { pages } from "@/content/pages";
@@ -21,25 +25,17 @@ export async function generateMetadata({ params }: { params: InstanceParams }) {
 export default async function Page({ params }: { params: InstanceParams }) {
   console.log(params);
   return (
-    <Unauthorised
-      title="Coming Soon"
-      message="This page is still under development. How did you even find it?"
-    />
-  );
-  {
-    /* 
     <PanelWrapper className="flex flex-col items-center justify-center gap-5 pt-8">
-      // TODO: add to NoteCard after merging branches 
-      <p>Include note card here explaining the steps/what will happen?</p>
-     
+      <NoteCard>
+        Include note card here explaining the steps/what will happen?
+      </NoteCard>
       <SubHeading className="mt-3">
         Check users exist on Assessment System
       </SubHeading>
       <SubHeading className="mt-16">
         Create Projects on Assessment System
       </SubHeading>
-      <CreateProjectsSection /> 
-      </PanelWrapper>
-      */
-  }
+      <CreateProjectsSection />
+    </PanelWrapper>
+  );
 }
